- various "cosmetics" cleanups to shut up gcc complaints on higher warning   levels (Ralf)

+++ methods/gzip.cc
@@ -94,7 +94,7 @@ bool GzipMethod::Fetch(FetchItem *Itm)
    while (1) 
    {
       unsigned char Buffer[4*1024];
-      unsigned long Count;
+      ssize_t Count;
       
       Count = read(GzOut[0],Buffer,sizeof(Buffer));
       if (Count < 0 && errno == EINTR)
